Abstractions for initializing upload property options.
