+++ client/galaxy/scripts/components/Upload/index.js
@@ -2,3 +2,5 @@
  * External entry point for Upload components (currently only the modal).
  */
 export { default as UploadModal } from "./UploadModal";
+
+export { initializeUploadDefaults } from "./config";
